ai-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1

# OCR
pytesseract==0.3.10
Pillow==10.3.0
opencv-python-headless==4.9.0.80

# AI / LLM
openai==1.25.1
httpx==0.27.0

# Message Queue
aio-pika==9.4.1      # RabbitMQ async client

# HTTP Client (callback to .NET API)
httpx==0.27.0

# Utilities
python-dotenv==1.0.1
structlog==24.1.0
